refactor: CreateItem service refactor
